refactor(context_compressor): extract content/media strip helpers (LB4)
Window 446-668 (banner-inclusive; pin 870-1111 minus leaves) into
context_compressor_media.py. 9 members re-exported is-identical through
the godfile; deps on typing/json/turn_context only (one-way). 43/43
targeted tests green.

Part of #78645
Part of #78647
